+++ src/core/createUnistylesImageBackground.tsx
@@ -1,59 +1,24 @@
 import React from 'react'
-import { getClassName } from './getClassname'
-import type { UnistylesValues } from '../types'
-import { isServer } from '../web/utils'
+import type { ImageBackground, ImageBackgroundProps } from 'react-native'
+import { UnistylesShadowRegistry } from '../specs'
+import { passForwardedRef } from './passForwardRef'
 
-type ComponentProps = {
-    style?: UnistylesValues | Array<UnistylesValues>
-}
+export const createUnistylesImageBackground = (Component: typeof ImageBackground) => React.forwardRef<ImageBackground, ImageBackgroundProps>((props, forwardedRef) => (
+    <Component
+        {...props}
+        ref={ref => passForwardedRef(props, ref, forwardedRef)}
+        imageRef={ref => {
+            const style = Array.isArray(props.imageStyle)
+                ? props.imageStyle
+                : [props.imageStyle]
 
-export const createUnistylesImageBackground = (Component: any) => React.forwardRef<HTMLElement, ComponentProps>((props, forwardedRef) => {
-    let storedRef: HTMLElement | null = null
-    let storedImageRef: HTMLElement | null = null
-    const classNames = getClassName(props.style)
-    const imageClassNames = getClassName(props.style)
+            // @ts-expect-error web types are not compatible with RN styles
+            UnistylesShadowRegistry.add(ref, style)
 
-    return (
-        <Component
-            {...props}
-            style={classNames}
-            imageStyle={imageClassNames}
-            imageRef={isServer() ? undefined : (ref: HTMLElement | null) => {
-                if (!ref) {
-                    // @ts-expect-error hidden from TS
-                    UnistylesShadowRegistry.remove(storedImageRef, imageClassNames?.hash)
-                }
-
-                storedImageRef = ref
+            return () => {
                 // @ts-expect-error hidden from TS
-                UnistylesShadowRegistry.add(ref, imageClassNames?.hash)
-
-                if (typeof forwardedRef === 'function') {
-                    return forwardedRef(ref)
-                }
-
-                if (forwardedRef) {
-                    forwardedRef.current = ref
-                }
-            }}
-            ref={isServer() ? undefined : (ref: HTMLElement | null) => {
-                if (!ref) {
-                    // @ts-expect-error hidden from TS
-                    UnistylesShadowRegistry.remove(storedRef, classNames?.hash)
-                }
-
-                storedRef = ref
-                // @ts-expect-error hidden from TS
-                UnistylesShadowRegistry.add(ref, classNames?.hash)
-
-                if (typeof forwardedRef === 'function') {
-                    return forwardedRef(ref)
-                }
-
-                if (forwardedRef) {
-                    forwardedRef.current = ref
-                }
-            }}
-        />
-    )
-})
+                UnistylesShadowRegistry.remove(ref)
+            }
+        }}
+    />
+))
